Try to improve numba call

See merge request benjello/OpenFisca-France-Pension!18

+++ openfisca_france_pension/tests/test_tools.py
@@ -8,7 +8,7 @@ from openfisca_france_pension.tools import mean_over_k_nonzero_largest
 
 
 def test_mean_over_k_nonzero_largest():
-    array = np.array([100] * 5 + [1000] * 5)
+    array = np.array([100] * 5 + [1000] * 5).astype(np.float32)
     assert mean_over_k_nonzero_largest(array, 4) == 1000
     assert mean_over_k_nonzero_largest(array, 5) == 1000
     assert mean_over_k_nonzero_largest(array, 6) == 850
@@ -16,7 +16,7 @@ def test_mean_over_k_nonzero_largest():
     assert mean_over_k_nonzero_largest(array, 10) == 550
     assert mean_over_k_nonzero_largest(array, 15) == 550
 
-    array = np.array([0] * 5 + [1000] * 5)
+    array = np.array([0] * 5 + [1000] * 5).astype(np.float32)
     assert mean_over_k_nonzero_largest(array, 4) == 1000
     assert mean_over_k_nonzero_largest(array, 5) == 1000
     assert mean_over_k_nonzero_largest(array, 6) == 1000
@@ -24,7 +24,7 @@ def test_mean_over_k_nonzero_largest():
     assert mean_over_k_nonzero_largest(array, 10) == 1000
     assert mean_over_k_nonzero_largest(array, 15) == 1000
 
-    array = np.array([0] * 5 + [100] * 5 + [1000] * 5)
+    array = np.array([0] * 5 + [100] * 5 + [1000] * 5).astype(np.float32)
     assert mean_over_k_nonzero_largest(array, 4) == 1000
     assert mean_over_k_nonzero_largest(array, 5) == 1000
     assert mean_over_k_nonzero_largest(array, 6) == 850
